Dojo/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-groq>=0.0.1
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.4.22
pypdf>=3.17.1
python-dateutil>=2.8.2
langchain
langgraph
langchain-groq
langchain-community
langchain-core
google-search-results
beautifulsoup4
gradio
streamlit
python-dotenv
requests
sentence-transformers
tf-keras
langchain-huggingface
torch 
accelerate 